requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
numpy>=2.0
pandas>=2.2.3
plotly==5.18.0
kaleido==0.2.1
scikit-learn>=1.4.2
supabase==2.4.6
razorpay==1.4.1
python-multipart==0.0.9
pytesseract==0.3.10
Pillow>=11.0.0
python-dotenv==1.0.1
pydantic>=2.6.1